+++ src/main/java/com/supermartijn642/movingelevators/gui/PlusMinusButtonWidget.java
@@ -1,6 +1,6 @@
 package com.supermartijn642.movingelevators.gui;
 
-import com.supermartijn642.core.gui.ScreenUtils;
+import com.supermartijn642.core.gui.GuiGraphicsHelper;
 import com.supermartijn642.core.gui.widget.WidgetRenderContext;
 import com.supermartijn642.core.gui.widget.premade.AbstractButtonWidget;
 import net.minecraft.network.chat.Component;
@@ -14,7 +14,7 @@ import java.util.function.Supplier;
  */
 public class PlusMinusButtonWidget extends AbstractButtonWidget {
 
-    private static final ResourceLocation PLUS_MINUS_BUTTONS = ResourceLocation.fromNamespaceAndPath("movingelevators", "textures/gui/plus_minus_buttons.png");
+    public static final ResourceLocation PLUS_MINUS_BUTTONS = ResourceLocation.fromNamespaceAndPath("movingelevators", "gui/plus_minus_buttons");
 
     private final boolean isPlus;
     private final Component hoverText;
@@ -40,8 +40,8 @@ public class PlusMinusButtonWidget extends AbstractButtonWidget {
     }
 
     @Override
-    public void render(WidgetRenderContext context, int mouseX, int mouseY){
+    public void render(WidgetRenderContext context, GuiGraphicsHelper graphics, int mouseX, int mouseY){
         this.active = this.isActive.get();
-        ScreenUtils.drawTexture(PLUS_MINUS_BUTTONS, context.poseStack(), this.x, this.y, this.width, this.height, this.isPlus ? 0 : 1 / 2f, this.active ? this.isFocused() ? 1 / 3f : 0 : 2 / 3f, 1 / 2f, 1 / 3f);
+        graphics.submitSprite(PLUS_MINUS_BUTTONS, this.x, this.y, this.width, this.height, p -> p.uv(this.isPlus ? 0 : 1 / 2f, this.active ? this.isFocused() ? 1 / 3f : 0 : 2 / 3f, 1 / 2f, 1 / 3f));
     }
 }