+++ changelog.md
@@ -1,2 +1,2 @@
 ### Moving Elevators 1.4.11
-- Initial release of Moving Elevators for Minecraft 1.21.5
+- Initial release of Moving Elevators for Minecraft 1.21.6 & 1.21.7